--- src/main/java/de/relluem94/minecraft/server/spigot/essentials/enchantment/EnchantLevel.java
@@ -1,3 +1,0 @@
-package de.relluem94.minecraft.server.spigot.essentials.enchantment;
-
-public record EnchantLevel(int startLevel,int maxLevel) {}
