--- src/main/java/de/relluem94/minecraft/server/spigot/essentials/enchantment/EnchantName.java
@@ -1,3 +1,0 @@
-package de.relluem94.minecraft.server.spigot.essentials.enchantment;
-
-public record EnchantName(String name, String displayName) {}
