+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/enums/EntityCoins.java
@@ -1,4 +1,4 @@
-package de.relluem94.minecraft.server.spigot.essentials.constants;
+package de.relluem94.minecraft.server.spigot.essentials.enums;
 
 import lombok.Getter;
 import org.bukkit.entity.EntityType;
