+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/events/protect/EntityBreakDoorProtect.java
@@ -2,7 +2,7 @@ package de.relluem94.minecraft.server.spigot.essentials.events.protect;
 
 import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.ProtectionHelper;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.ProtectionEntry;
+import de.relluem94.minecraft.server.spigot.essentials.model.pojo.ProtectionEntry;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.event.EventHandler;
