+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/model/enchantment/EnchantLevel.java
@@ -1,0 +1,3 @@
+package de.relluem94.minecraft.server.spigot.essentials.model.enchantment;
+
+public record EnchantLevel(int startLevel,int maxLevel) {}
