+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/model/enchantment/EnchantName.java
@@ -1,0 +1,3 @@
+package de.relluem94.minecraft.server.spigot.essentials.model.enchantment;
+
+public record EnchantName(String name, String displayName) {}
