+++ src/test/java/de/relluem94/minecraft/server/spigot/essentials/enums/EntityCoinsTest.java
@@ -1,4 +1,4 @@
-package de.relluem94.minecraft.server.spigot.essentials.constants;
+package de.relluem94.minecraft.server.spigot.essentials.enums;
 
 import org.bukkit.entity.EntityType;
 import org.junit.jupiter.api.Test;
